Wrap to ShowMessage for options on Unix :anchor:
- Fix default text and title labels
