+++ githooks/apps/dialog/gui/extraButtons.go
@@ -1,4 +1,4 @@
-// +build darwin unix
+// +build !windows
 
 package gui
